refactor: Separate context
